New PIC16F83 blinking fast example

+++ ports/Microchip/PIC12F-16F-generic/PIC16F-core/api/pin/high_fast.c.v
@@ -8,10 +8,10 @@ module pin
 
 #include "high_fast.c"
 
-fn C.HIGH_FAST(port_name u8, pin_number u8)
+fn C.HIGH_FAST(port_name int, pin_number int)
 
 // high_fast puts a logic 1 to a pin faster than the high function
 @[as_macro]
-pub fn high_fast(port_name u8, pin_number u8) {
+pub fn high_fast(port_name int, pin_number int) {
 	C.HIGH_FAST(port_name, pin_number)
 }
